project update

+++ books_recommender/components/stage_00_data_ingestion.py
@@ -27,15 +27,19 @@ class DataIngestion:
         
         """
         try:
-            
             dataset_url = self.data_ingestion_config.dataset_download_url
             zip_download_dir = self.data_ingestion_config.raw_data_dir
             os.makedirs(zip_download_dir, exist_ok=True)
             data_file_name = os.path.basename(dataset_url)
             zip_file_path = os.path.join(zip_download_dir, data_file_name)
-            logging.info(f"Downloading data from {dataset_url} into file {zip_file_path}")
-            urllib.request.urlretrieve(dataset_url,zip_file_path)
-            logging.info(f"Downloaded data from {dataset_url} into file {zip_file_path}")
+
+            if not os.path.exists(zip_file_path):
+                logging.info(f"Downloading data from {dataset_url} into file {zip_file_path}")
+                urllib.request.urlretrieve(dataset_url, zip_file_path)
+                logging.info(f"Downloaded data from {dataset_url} into file {zip_file_path}")
+            else:
+                logging.info(f"File {zip_file_path} already exists. Skipping download.")
+
             return zip_file_path
 
         except Exception as e:
@@ -51,9 +55,10 @@ class DataIngestion:
         try:
             ingested_dir = self.data_ingestion_config.ingested_dir
             os.makedirs(ingested_dir, exist_ok=True)
+            logging.info(f"Extracting zip file: {zip_file_path} into dir: {ingested_dir}")
             with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
                 zip_ref.extractall(ingested_dir)
-            logging.info(f"Extracting zip file: {zip_file_path} into dir: {ingested_dir}")
+                logging.info(f"Extraction complete. Files extracted: {zip_ref.namelist()}")
         except Exception as e:
             raise AppException(e,sys) from e
